+++ src/Gargantext/Core/Text/Corpus/API.hs
@@ -18,6 +18,8 @@ module Gargantext.Core.Text.Corpus.API
   )
     where
 
+import Conduit
+import Data.Either (Either(..))
 import Data.Maybe
 import Gargantext.API.Admin.Orchestrator.Types (ExternalAPIs(..), externalAPIs)
 import Gargantext.Core (Lang(..))
@@ -27,21 +29,25 @@ import qualified Gargantext.Core.Text.Corpus.API.Hal     as HAL
 import qualified Gargantext.Core.Text.Corpus.API.Isidore as ISIDORE
 import qualified Gargantext.Core.Text.Corpus.API.Istex   as ISTEX
 import qualified Gargantext.Core.Text.Corpus.API.Pubmed  as PUBMED
-
--- | TODO put in gargantext.init
-default_limit :: Maybe Integer
-default_limit = Just 10000
+import Servant.Client (ClientError)
 
 -- | Get External API metadata main function
 get :: ExternalAPIs
     -> Lang
     -> Query
     -> Maybe Limit
-    -> IO [HyperdataDocument]
-get PubMed  _la q _l = PUBMED.get   q default_limit -- EN only by default
-get HAL      la q _l = HAL.get   la q default_limit
-get IsTex    la q _l = ISTEX.get la q default_limit
-get Isidore  la q _l = ISIDORE.get la (fromIntegral <$> default_limit) (Just q) Nothing
+    -- -> IO [HyperdataDocument]
+    -> IO (Either ClientError (Maybe Integer, ConduitT () HyperdataDocument IO ()))
+get PubMed  _la q limit = PUBMED.get q limit
+  --docs <- PUBMED.get   q default_limit -- EN only by default
+  --pure (Just $ fromIntegral $ length docs, yieldMany docs)
+get HAL      la q limit = HAL.getC  la q limit
+get IsTex    la q limit = do
+  docs <- ISTEX.get la q limit
+  pure $ Right (Just $ fromIntegral $ length docs, yieldMany docs)
+get Isidore  la q limit = do
+  docs <- ISIDORE.get la (fromIntegral <$> limit) (Just q) Nothing
+  pure $ Right (Just $ fromIntegral $ length docs, yieldMany docs)
 get _        _  _ _ = undefined
 
 -- | Some Sugar for the documentation